% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{bulker}
\alias{bulker}
\title{Report bulker}
\usage{
bulker(inputString)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Report bulker- looks at what is in the report, then adds another row's report to the current to bulk it up.
It specifically doesn't add reports that mention the same compartment to avoid contradictions
}
\examples{
bulker()
}
\keyword{cats}
