% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{flag}
\alias{flag}
\title{flag}
\usage{
flag(inputString, outputString, proportion)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
flag
}
\examples{
flag("Normal gastroscopy to the duodenum.","Hiatus",8)
}
\keyword{cats}
