% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{IntRanNumber}
\alias{IntRanNumber}
\title{Always interpolate- random number of elements from a list}
\usage{
IntRanNumber(inputString, listtoSample)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
IntRanNumber("exudates|Pull|crepe|Rings|tight",FD_EosinophilicDetail)
}
\keyword{cats}
