% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{IntRanOneElement}
\alias{IntRanOneElement}
\title{Always interpolate- random 1 element from a list}
\usage{
IntRanOneElement(inputString, listtoSample)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This adds an extra phrase from a list based on flag- check to see if IntRanOneElement1 already does this
}
\examples{
IntRanOneElement()
}
\keyword{cats}
