% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{IntRanOneElement1}
\alias{IntRanOneElement1}
\title{Internally relevant detail adder - interpolate}
\usage{
IntRanOneElement1(inputString, listtoSample)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This adds the detail to a report based on the input string (ie based on the flag)
}
\examples{
IntRanOneElement1()
}
\keyword{cats}
