% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{IntRanOneElementWithPhrase}
\alias{IntRanOneElementWithPhrase}
\title{Internally relevant detail adder with extra phrase- interpolate}
\usage{
IntRanOneElementWithPhrase(inputString, listtoSample, addedPhrase1,
  addedPhrase2)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
IntRanOneElementWithPhrase()
}
\keyword{cats}
