% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{listtodf}
\alias{listtodf}
\title{This function allows you to express your love of cats.}
\usage{
listtodf(thelist)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\keyword{cats}
