% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMaker.R
\name{RandomSingleGsub}
\alias{RandomSingleGsub}
\title{RandomSingleGsub}
\usage{
RandomSingleGsub(phraseToReplace, listOfReplacements)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
RandomSingleGsub()
}
\keyword{cats}
